update names, remove extra comments, add documentation
